YP-35: Moving all DTOs to the response package
