--- backend/src/main/java/com/portfolio/dto/BusinessDto.java
@@ -1,13 +1,0 @@
-package com.portfolio.dto;
-
-import lombok.Data;
-
-@Data
-public class BusinessDto {
-
-    private Integer id;
-    private String name;
-    private String logoUrl;
-    private AddressDto addressDto;
-
-}
