+++ backend/src/main/java/com/portfolio/dto/response/BusinessDto.java
@@ -1,14 +1,15 @@
-package com.portfolio.dto;
+package com.portfolio.dto.response;
 
 import lombok.Data;
 
 import java.io.Serializable;
 
 @Data
-public class TechnologyDto implements Serializable {
+public class BusinessDto implements Serializable {
 
     private Integer id;
     private String name;
     private String logoUrl;
+    private AddressDto addressDto;
 
 }
