+++ backend/src/main/java/com/portfolio/dto/response/CareerDto.java
@@ -1,4 +1,4 @@
-package com.portfolio.dto;
+package com.portfolio.dto.response;
 
 import lombok.Data;
