+++ backend/src/main/java/com/portfolio/dto/response/InstituteUniversityDto.java
@@ -1,9 +1,11 @@
-package com.portfolio.dto;
+package com.portfolio.dto.response;
 
 import lombok.Data;
 
+import java.io.Serializable;
+
 @Data
-public class InstituteUniversityDto {
+public class InstituteUniversityDto implements Serializable {
 
     private Integer id;
     private String name;
