+++ backend/src/main/java/com/portfolio/dto/response/TechnologyDto.java
@@ -1,0 +1,14 @@
+package com.portfolio.dto.response;
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+@Data
+public class TechnologyDto implements Serializable {
+
+    private Integer id;
+    private String name;
+    private String logoUrl;
+
+}
