+++ backend/src/main/java/com/portfolio/mapper/ICourseMapper.java
@@ -1,6 +1,6 @@
 package com.portfolio.mapper;
 
-import com.portfolio.dto.CourseDto;
+import com.portfolio.dto.response.CourseDto;
 import com.portfolio.model.entity.Course;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
