+++ backend/src/main/java/com/portfolio/mapper/IJobExperienceMapper.java
@@ -1,6 +1,6 @@
 package com.portfolio.mapper;
 
-import com.portfolio.dto.JobExperienceDto;
+import com.portfolio.dto.response.JobExperienceDto;
 import com.portfolio.model.entity.JobExperience;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
