+++ backend/src/main/java/com/portfolio/mapper/IPersonMapper.java
@@ -1,6 +1,6 @@
 package com.portfolio.mapper;
 
-import com.portfolio.dto.PersonDto;
+import com.portfolio.dto.response.PersonDto;
 import com.portfolio.model.entity.Person;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
