+++ backend/src/main/java/com/portfolio/service/IBusinessService.java
@@ -1,6 +1,6 @@
 package com.portfolio.service;
 
-import com.portfolio.dto.BusinessDto;
+import com.portfolio.dto.response.BusinessDto;
 
 import java.util.List;
