+++ backend/src/main/java/com/portfolio/service/impl/PersonServiceImpl.java
@@ -1,6 +1,6 @@
 package com.portfolio.service.impl;
 
-import com.portfolio.dto.PersonDto;
+import com.portfolio.dto.response.PersonDto;
 import com.portfolio.mapper.IPersonMapper;
 import com.portfolio.model.entity.Person;
 import com.portfolio.repository.IPersonRepository;
